+++ R/vcr-package.R
@@ -30,9 +30,11 @@
 
 ## usethis namespace: start
 #' @import rlang
-#' @importFrom R6 R6Class
-#' @importFrom utils getParseData
-#' @importFrom yaml yaml.load yaml.load_file as.yaml
-#' @importFrom urltools url_parse url_compose
 ## usethis namespace: end
 NULL
+
+# https://r-pkgs.org/dependencies-in-practice.html#how-to-not-use-a-package-in-imports
+ignore_unused_imports <- function() {
+  yaml::read_yaml
+  R6::R6Class
+}