+++ NEWS.md
@@ -4,6 +4,7 @@ vcr (development version)
 ### BREAKING CHANGES
 
 * Internal `real_http_connections_allowed()` is no longer exported and has been removed.
+* Internal `Request` and `VcrResponse` classes are no longer exported and have been removed.
 * `HTTPInteractionList` is no longer exported; it's an internal implementation detail.
 * The `uri_parser` option is no longer supported.
 * `vcr_configuration(allow_http_connections_when_no_cassette)` is no longer supported. It hasn't worked for a while.